requirements.txt
lightrag-hku
ragas>=0.4.0
streamlit
pypdf
sentence-transformers
openai
python-dotenv
plotly
nest-asyncio
numpy
pytest
pytest-asyncio